+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Breaking Changes 🛠
+
+- Renamed `RwSignalResource` to `RwSignalSynced`
+- Renamed `signal_resource` to `signal_synced`
+
+### New Features 🎉
+
+- `RwSignalSynced` now implements `Write` and `Copy` + `Clone`
+- query...
+
+
 ## [0.1.0] - 2024-12-03
 
 - Updated to Leptos 0.7.0 and Bevy 0.15.0